[arduino_tape_emulator] Change type from size_t to uint16_t

+++ arduino_tape_emulator/BlockHandler.h
@@ -20,9 +20,9 @@ class BlockHandler
 {
 private:
     volatile byte*      mBufferIn;
-    volatile size_t     mLengthIn;
+    volatile uint16_t   mLengthIn;
     volatile byte*      mBufferOut;
-    volatile size_t     mLengthOut;
+    volatile uint16_t   mLengthOut;
 
     enum STAGE
     {
@@ -36,19 +36,19 @@ private:
         FINISH
     };
 
-    size_t      mIndexByte;
+    uint16_t    mIndexByte;
     byte        mMask;
     byte        mCurrentByte;
 
     volatile STAGE  mStage;
     bool            mCurrentBitOne;
     bool            mMeanderUp;
-    size_t          mPeriod;
+    uint16_t        mPeriod;
 
-    size_t mImpulseCouter;
+    uint16_t        mImpulseCouter;
 
 public:
-    BlockHandler(size_t bufferSize) : mStage(STAGE::FINISH)
+    BlockHandler(uint16_t bufferSize) : mStage(STAGE::FINISH)
     {
         mBufferIn = new byte[bufferSize];
         mBufferOut = new byte[bufferSize];
@@ -66,7 +66,7 @@ public:
         return 0 == mLengthIn;
     }
 
-    byte* fillBuffer(byte *buffer, size_t length)
+    byte* fillBuffer(byte *buffer, uint16_t length)
     {
         auto old_buffer = mBufferIn;
         mBufferIn = buffer;
